chore(transparent-proxy): clean and improve tproxy config source code
Some files currently have more than 1000 lines, which makes it harder to
maintain, so I made a decision to move helper funcions to separate file
called `utils.go` (I don't like having utils files, but I belive it's a
better decision than having huge files)
