+++ pkg/transparentproxy/config/config.go
@@ -7,7 +7,6 @@ import (
 	"io"
 	"net"
 	"os"
-	"os/user"
 	"slices"
 	"strconv"
 	"strings"
@@ -15,7 +14,6 @@ import (
 
 	"github.com/miekg/dns"
 	"github.com/pkg/errors"
-	"github.com/vishvananda/netlink"
 
 	core_config "github.com/kumahq/kuma/pkg/config"
 	config_types "github.com/kumahq/kuma/pkg/config/types"
@@ -181,16 +179,6 @@ func (p *Ports) Set(s string) error {
 	}
 
 	return nil
-}
-
-func parsePort(s string) (Port, error) {
-	u, err := parseUint16(s)
-
-	if err != nil || u == 0 {
-		return 0, errors.Errorf("value '%s' is not a valid port (uint16 in the range [1, 65535])", s)
-	}
-
-	return Port(u), nil
 }
 
 func (p *Ports) UnmarshalJSON(bs []byte) error {
@@ -918,281 +906,4 @@ func DefaultConfig() Config {
 		StoreFirewalld: false,
 		CNIMode:        false,
 	}
-}
-
-// getLoopbackInterfaceName retrieves the name of the loopback interface on the
-// system. This function iterates over all network interfaces and checks if the
-// 'net.FlagLoopback' flag is set. If a loopback interface is found, its name is
-// returned. Otherwise, an error message indicating that no loopback interface
-// was found is returned.
-func getLoopbackInterfaceName() (string, error) {
-	interfaces, err := net.Interfaces()
-	if err != nil {
-		return "", errors.Wrap(err, "failed to retrieve network interfaces")
-	}
-
-	for _, iface := range interfaces {
-		if iface.Flags&net.FlagLoopback != 0 {
-			return iface.Name, nil
-		}
-	}
-
-	return "", errors.New("no loopback interface found on the system")
-}
-
-// parseExcludePortsForUIDs parses a slice of strings representing port
-// exclusion rules based on UIDs and returns a slice of Exclusion structs.
-//
-// Each input string should follow the format: <protocol:>?<ports:>?<uids>.
-// This means the string can contain optional protocol and port values,
-// followed by mandatory UID values. Examples of valid formats include:
-//   - "tcp:22:1000-2000" (TCP protocol, port 22, UIDs from 1000 to 2000)
-//   - "udp:53:1001" (UDP protocol, port 53, UID 1001)
-//   - "80:1002" (Any protocol, port 80, UID 1002)
-//   - "1003" (Any protocol, any port, UID 1003)
-func parseExcludePortsForUIDs(exclusionRules []string) ([]Exclusion, error) {
-	var result []Exclusion
-
-	for _, elem := range exclusionRules {
-		parts := strings.Split(elem, ":")
-		if len(parts) == 0 || len(parts) > 3 {
-			return nil, errors.Errorf(
-				"invalid format for excluding ports by UIDs: '%s'. Expected format: <protocol:>?<ports:>?<uids>",
-				elem,
-			)
-		}
-
-		var portValuesOrRange, protocolOpts, uidValuesOrRange string
-
-		switch len(parts) {
-		case 1:
-			protocolOpts = "*"
-			portValuesOrRange = "*"
-			uidValuesOrRange = parts[0]
-		case 2:
-			protocolOpts = "*"
-			portValuesOrRange = parts[0]
-			uidValuesOrRange = parts[1]
-		case 3:
-			protocolOpts = parts[0]
-			portValuesOrRange = parts[1]
-			uidValuesOrRange = parts[2]
-		}
-
-		if uidValuesOrRange == "*" {
-			return nil, errors.New("wildcard '*' is not allowed for UIDs")
-		}
-
-		if portValuesOrRange == "*" || portValuesOrRange == "" {
-			portValuesOrRange = "1-65535"
-		}
-
-		if err := validateUintValueOrRange(portValuesOrRange); err != nil {
-			return nil, errors.Wrap(err, "invalid port range")
-		}
-
-		if strings.Contains(uidValuesOrRange, ",") {
-			return nil, errors.Errorf(
-				"invalid UID entry: '%s'. It should either be a single item or a range",
-				uidValuesOrRange,
-			)
-		}
-
-		if err := validateUintValueOrRange(uidValuesOrRange); err != nil {
-			return nil, errors.Wrap(err, "invalid UID range")
-		}
-
-		var protocols []ProtocolL4
-		if protocolOpts == "" || protocolOpts == "*" {
-			protocols = []ProtocolL4{ProtocolTCP, ProtocolUDP}
-		} else {
-			for _, s := range strings.Split(protocolOpts, ",") {
-				if p := ParseProtocolL4(s); p != ProtocolUndefined {
-					protocols = append(protocols, p)
-					continue
-				}
-
-				return nil, errors.Errorf(
-					"invalid or unsupported protocol: '%s'",
-					s,
-				)
-			}
-		}
-
-		for _, p := range protocols {
-			ports := strings.ReplaceAll(portValuesOrRange, "-", ":")
-			uids := strings.ReplaceAll(uidValuesOrRange, "-", ":")
-
-			result = append(result, Exclusion{
-				Ports:    ValueOrRangeList(ports),
-				UIDs:     ValueOrRangeList(uids),
-				Protocol: p,
-			})
-		}
-	}
-
-	return result, nil
-}
-
-// parseExcludePortsForIPs parses a slice of strings representing port exclusion
-// rules based on IP addresses and returns a slice of IPToPorts structs.
-//
-// This function currently allows each exclusion rule to be a valid IPv4 or IPv6
-// address, with or without a CIDR suffix. It is designed to potentially support
-// more complex exclusion rules in the future.
-func parseExcludePortsForIPs(
-	exclusionRules []string,
-	ipv6 bool,
-) ([]Exclusion, error) {
-	var result []Exclusion
-
-	for _, rule := range exclusionRules {
-		if rule == "" {
-			return nil, errors.New(
-				"invalid exclusion rule: the rule cannot be empty",
-			)
-		}
-
-		for _, address := range strings.Split(rule, ",") {
-			err, isExpectedIPVersion := validateIP(address, ipv6)
-			if err != nil {
-				return nil, errors.Wrap(err, "invalid exclusion rule")
-			}
-
-			if isExpectedIPVersion {
-				result = append(result, Exclusion{Address: address})
-			}
-		}
-	}
-
-	return result, nil
-}
-
-// validateUintValueOrRange validates whether a given string represents a valid
-// single uint16 value or a range of uint16 values. The input string can contain
-// multiple comma-separated values or ranges (e.g., "80,1000-2000").
-func validateUintValueOrRange(valueOrRange string) error {
-	for _, element := range strings.Split(valueOrRange, ",") {
-		for _, port := range strings.Split(element, "-") {
-			if _, err := parseUint16(port); err != nil {
-				return errors.Wrapf(
-					err,
-					"validation failed for value or range '%s'",
-					valueOrRange,
-				)
-			}
-		}
-	}
-
-	return nil
-}
-
-// validateIP validates an IP address or CIDR and checks if it matches the
-// expected IP version (IPv4 or IPv6).
-func validateIP(address string, ipv6 bool) (error, bool) {
-	// Attempt to parse the address as a CIDR.
-	ip, _, err := net.ParseCIDR(address)
-	// If parsing as CIDR fails, attempt to parse it as a plain IP address.
-	if err != nil {
-		ip = net.ParseIP(address)
-	}
-
-	// If parsing as both CIDR and IP address fails, return an error with a
-	// message.
-	if ip == nil {
-		return errors.Errorf(
-			"invalid IP address: '%s'. Expected format: <ip> or <ip>/<cidr> (e.g., 10.0.0.1, 172.16.0.0/16, fe80::1, fe80::/10)",
-			address,
-		), false
-	}
-
-	// Check if the IP version matches the expected IP version.
-	// For IPv4, ip.To4() will not be nil. For IPv6, ip.To4() will be nil.
-	return nil, ipv6 == (ip.To4() == nil)
-}
-
-// parseUint16 parses a string representing a uint16 value and returns its
-// uint16 representation.
-func parseUint16(port string) (uint16, error) {
-	parsedPort, err := strconv.ParseUint(port, 10, 16)
-	if err != nil {
-		return 0, errors.Errorf("invalid uint16 value: '%s'", port)
-	}
-
-	return uint16(parsedPort), nil
-}
-
-// hasLocalIPv6 checks if the local system has an active non-loopback IPv6
-// address. It scans through all network interfaces to find any IPv6 address
-// that is not a loopback address.
-func hasLocalIPv6() (bool, error) {
-	addrs, err := net.InterfaceAddrs()
-	if err != nil {
-		return false, err
-	}
-
-	for _, a := range addrs {
-		if ipnet, ok := a.(*net.IPNet); ok &&
-			!ipnet.IP.IsLoopback() &&
-			ipnet.IP.To4() == nil {
-			return true, nil
-		}
-	}
-
-	return false, errors.New("no local IPv6 addresses detected")
-}
-
-// configureIPv6OutboundAddress sets up a dedicated IPv6 address (::6) on the
-// loopback interface ("lo") for our transparent proxy functionality.
-//
-// Background:
-//   - The default IPv6 configuration (prefix length 128) only allows binding to
-//     the loopback address (::1).
-//   - Our transparent proxy requires a distinct IPv6 address (::6 in this case)
-//     to identify traffic processed by the kuma-dp sidecar.
-//   - This identification allows for further processing and avoids redirection
-//     loops.
-//
-// This function is equivalent to running the command:
-// `ip -6 addr add "::6/128" dev lo`
-func configureIPv6OutboundAddress() error {
-	link, err := netlink.LinkByName("lo")
-	if err != nil {
-		return errors.Wrap(err, "failed to find loopback interface ('lo')")
-	}
-
-	// Equivalent to "::6/128"
-	addr := &netlink.Addr{
-		IPNet: &net.IPNet{
-			IP:   net.ParseIP("::6"),
-			Mask: net.CIDRMask(128, 128),
-		},
-	}
-
-	if err := netlink.AddrAdd(link, addr); err != nil {
-		// Address already exists, ignore error and continue
-		if strings.Contains(strings.ToLower(err.Error()), "file exists") {
-			return nil
-		}
-
-		return errors.Wrapf(err, "failed to add IPv6 address %s to loopback interface", addr.IPNet)
-	}
-
-	return nil
-}
-
-func findUserUID(userOrUID string) (string, bool) {
-	if userOrUID == "" {
-		return "", false
-	}
-
-	if u, err := user.LookupId(userOrUID); err == nil {
-		return u.Uid, true
-	}
-
-	if u, err := user.Lookup(userOrUID); err == nil {
-		return u.Uid, true
-	}
-
-	return "", false
 }
