+++ pkg/transparentproxy/config/config_executables.go
@@ -1,14 +1,12 @@
 package config
 
 import (
-	"bufio"
 	"bytes"
 	"context"
 	std_errors "errors"
 	"fmt"
 	"os"
 	"os/exec"
-	"path/filepath"
 	"slices"
 	"strings"
 	"time"
@@ -464,207 +462,4 @@ func (c Executables) InitializeIPv4(
 	default:
 		return nft, c.NftIPv6, nil
 	}
-}
-
-// buildRestoreArgs constructs a slice of flags for restoring iptables rules
-// based on the provided configuration and iptables mode.
-//
-// This function generates a list of command-line flags to be used with
-// iptables-restore, tailored to the given parameters:
-//   - For non-legacy iptables mode, it returns an empty slice, as no additional
-//     flags are required.
-//   - For legacy mode, it conditionally adds the `--wait` and `--wait-interval`
-//     flags based on the provided configuration values.
-func buildRestoreArgs(cfg Config, mode IptablesMode) []string {
-	var flags []string
-
-	if mode != IptablesModeLegacy {
-		return flags
-	}
-
-	if cfg.Wait > 0 {
-		flags = append(flags, fmt.Sprintf("%s=%d", FlagWait, cfg.Wait))
-	}
-
-	if cfg.WaitInterval > 0 {
-		flags = append(flags, fmt.Sprintf(
-			"%s=%d",
-			FlagWaitInterval,
-			cfg.WaitInterval,
-		))
-	}
-
-	return flags
-}
-
-// getPathsToSearchForExecutable generates a list of potential paths for the
-// given executable considering both versions with and without the mode suffix.
-//
-// This function prioritizes finding the executable with the mode information
-// embedded in the name (e.g., iptables-nft) for faster mode verification.
-// It achieves this by:
-//  1. Adding the `nameWithMode` (e.g., iptables-nft) as the first potential
-//     path.
-//  2. Appending paths formed by joining `FallbackExecutablesSearchLocations`
-//     with `nameWithMode` (e.g., /usr/sbin/iptables-nft, /sbin/iptables-nft).
-//  3. After checking paths with the mode suffix, it adds the `nameWithoutMode`
-//     (e.g., iptables) as a fallback.
-//  4. Similar to step 2, it appends paths formed by joining
-//     `FallbackExecutablesSearchLocations` with `nameWithoutMode`.
-//
-// Finally, the function returns the combined list of potential paths for the
-// executable.
-func getPathsToSearchForExecutable(
-	nameWithMode string,
-	nameWithoutMode string,
-) []string {
-	var paths []string
-
-	paths = append(paths, nameWithMode)
-	for _, fallbackPath := range FallbackExecutablesSearchLocations {
-		paths = append(paths, filepath.Join(fallbackPath, nameWithMode))
-	}
-
-	paths = append(paths, nameWithoutMode)
-	for _, fallbackPath := range FallbackExecutablesSearchLocations {
-		paths = append(paths, filepath.Join(fallbackPath, nameWithoutMode))
-	}
-
-	return paths
-}
-
-// findPath attempts to locate the executable named by 'path' on the system.
-//
-// This function uses exec.LookPath to search for the executable based on the
-// following logic:
-//   - If 'path' contains a slash (/), it's considered an absolute path and
-//     searched for directly.
-//   - If 'path' doesn't contain a slash:
-//   - LookPath searches for the executable in directories listed in the
-//     system's PATH environment variable.
-//   - In Go versions before 1.19, a relative path to the current working
-//     directory could be returned for non-absolute paths. In Go 1.19 and
-//     later, such cases will result in an exec.ErrDot error with the relative
-//     path.
-//
-// The function handles these cases as follows:
-// - If no error occurs, the absolute path found by exec.LookPath is returned.
-// - If exec.ErrDot is encountered:
-//   - The current working directory is retrieved using os.Getwd().
-//     If successful:
-//   - The relative path found by exec.LookPath is prepended with the current
-//     working directory using filepath.Join to create an absolute path.
-//   - If getting the current working directory fails:
-//   - The original relative path found by LookPath is returned as a fallback
-//
-// If no path is found or an unexpected error occurs, an empty string is
-// returned.
-func findPath(path string) string {
-	found, err := exec.LookPath(path)
-	switch {
-	case err == nil:
-		return found
-	case errors.Is(err, exec.ErrDot):
-		// Go 1.19+ behavior: relative path found. Try to prepend the current
-		// working directory.
-		if pwd, err := os.Getwd(); err == nil {
-			return filepath.Join(pwd, found)
-		}
-
-		// Couldn't get the current working directory, fallback to the relative
-		// path.
-		return found
-	}
-
-	return ""
-}
-
-// joinNonEmptyWithHyphen joins a slice of strings with hyphens (-) as
-// separators, omitting any empty strings from the joined result.
-func joinNonEmptyWithHyphen(elems ...string) string {
-	return strings.Join(
-		slices.DeleteFunc(
-			elems,
-			func(s string) bool {
-				return s == ""
-			},
-		),
-		"-",
-	)
-}
-
-// createBackupFile generates a backup file with a specified prefix and a
-// timestamp suffix. The file is created in the system's temporary directory and
-// is used to store iptables rules for backup purposes.
-func createBackupFile(prefix string) (*os.File, error) {
-	// Generate a timestamp suffix for the backup file name.
-	dateSuffix := time.Now().Format("2006-01-02-150405")
-
-	// Construct the backup file name using the provided prefix and the
-	// timestamp suffix.
-	fileName := fmt.Sprintf("%s-rules.%s.txt.backup", prefix, dateSuffix)
-	filePath := filepath.Join(os.TempDir(), fileName)
-
-	// Create the backup file in the system's temporary directory.
-	f, err := os.Create(filePath)
-	if err != nil {
-		return nil, errors.Wrapf(err, "failed to create backup file: %s", filePath)
-	}
-
-	return f, nil
-}
-
-// createTempFile generates a temporary file with a specified prefix. The file
-// is created in the system's default temporary directory and is used for
-// storing iptables rules temporarily.
-//
-// This function performs the following steps:
-//  1. Constructs a template for the temporary file name using the provided
-//     prefix.
-//  2. Creates the temporary file in the system's default temporary directory.
-func createTempFile(prefix string) (*os.File, error) {
-	// Construct a template for the temporary file name using the provided prefix.
-	nameTemplate := fmt.Sprintf("%s-rules.*.txt", prefix)
-
-	// Create the temporary file in the system's default temporary directory.
-	f, err := os.CreateTemp(os.TempDir(), nameTemplate)
-	if err != nil {
-		return nil, errors.Wrapf(err, "failed to create temporary file with template: %s", nameTemplate)
-	}
-
-	return f, nil
-}
-
-// writeToFile writes the provided content to the specified file using a
-// buffered writer. It ensures that all data is written to the file by flushing
-// the buffer.
-//
-// This function performs the following steps:
-//  1. Writes the content to the file using a buffered writer for efficiency.
-//  2. Flushes the buffer to ensure all data is written to the file.
-func writeToFile(content string, f *os.File) error {
-	// Write the content to the file using a buffered writer.
-	writer := bufio.NewWriter(f)
-	if _, err := writer.WriteString(content); err != nil {
-		return errors.Wrapf(err, "failed to write to file: %s", f.Name())
-	}
-
-	// Flush the buffer to ensure all data is written.
-	if err := writer.Flush(); err != nil {
-		return errors.Wrapf(err, "failed to flush the buffered writer for file: %s", f.Name())
-	}
-
-	return nil
-}
-
-func handleRunError(err error, stderr *bytes.Buffer) error {
-	if stderr.Len() > 0 {
-		stderrTrimmed := strings.TrimSpace(stderr.String())
-		stderrLines := strings.Split(stderrTrimmed, "\n")
-		stderrFormated := strings.Join(stderrLines, ", ")
-
-		return errors.Errorf("%s: %s", err, stderrFormated)
-	}
-
-	return err
 }
